Function objc_sys::try_catch

pub unsafe extern "C" fn try_catch(
    f: extern "C" fn(_: *mut c_void),
    context: *mut c_void,
    error: *mut *mut objc_object,
) -> c_uchar

Call the given function inside an Objective-C @try/@catch block.

Defined in extern/exception.m and compiled in build.rs.

Alternatively, we could manually write assembly for this function like objrs does does, that would cut down on a build stage (and would probably give us a bit better performance), but it gets unwieldy very quickly, so I chose the much more stable option.

Another thing to remember: While Rust’s and Objective-C’s unwinding mechanisms are similar now, Rust’s is explicitly unspecified, and they may diverge significantly in the future; so handling this in pure Rust (using mechanisms like core::intrinsics::r#try) is not an option!