Crate objc_sys

Raw bindings to Objective-C runtimes

These bindings contain almost no documentation, so it is highly recommended to read the documentation of the original libraries:

Apple’s documentation about the Objective-C runtime.

Apple’s runtime reference.

Apple’s objc4 source code, in particular runtime.h.

GNUStep’s libobjc2 source code, in particular runtime.h.

See also the README.md for more background information, and for how to configure the desired runtime.

Structs

objc_class

An opaque type that represents an Objective-C class.

objc_ivar

An opaque type that represents an instance variable.

objc_method

A type that represents a method in a class definition.

objc_method_description

Describes an Objective-C method.

objc_object

An opaque type that represents an object / an instance of a class.

objc_property

An opaque type that describes a property in a class.

objc_property_attribute_t

Describes an Objective-C property attribute.

objc_protocol

Opaque type for Objective-C protocols.

objc_selector

An opaque type that represents a method selector.

objc_super

Specifies data used when sending messages to superclasses.

Constants

NO

The equivalent of false for Objective-C’s BOOL type.

NSIntegerMax

The maximum value for a NSInteger.

NSIntegerMin

The minimum value for a NSInteger.

NSUIntegerMax

The maximum value for a NSUInteger.

Nil

A quick alias for a null_mut class.

OBJC_ASSOCIATION_ASSIGN

Specifies a weak reference to the associated object.

OBJC_ASSOCIATION_COPY

Specifies that the associated object is copied.

OBJC_ASSOCIATION_COPY_NONATOMIC

Specifies that the associated object is copied.

OBJC_ASSOCIATION_RETAIN

Specifies a strong reference to the associated object.

OBJC_ASSOCIATION_RETAIN_NONATOMIC

Specifies a strong reference to the associated object.

OBJC_SYNC_NOT_INITIALIZED

Only relevant before macOS 10.13

OBJC_SYNC_NOT_OWNING_THREAD_ERROR

OBJC_SYNC_SUCCESS

OBJC_SYNC_TIMED_OUT

Only relevant before macOS 10.13

YES

The equivalent of true for Objective-C’s BOOL type.

nil

A quick alias for a null_mut object / instance.

Functions

_objc_msgForward⚠

_objc_msgForward_stret⚠

class_addIvar⚠

class_addMethod⚠

class_addProperty⚠

class_addProtocol⚠

class_conformsToProtocol⚠

class_copyIvarList⚠

class_copyMethodList⚠

class_copyPropertyList⚠

class_copyProtocolList⚠

class_createInstance⚠

class_getClassMethod⚠

class_getClassVariable⚠

class_getImageName⚠

class_getInstanceMethod⚠

class_getInstanceSize⚠

class_getInstanceVariable⚠

class_getIvarLayout⚠

class_getMethodImplementation⚠

class_getMethodImplementation_stret⚠

class_getName⚠

class_getProperty⚠

class_getSuperclass⚠

class_getVersion⚠

class_getWeakIvarLayout⚠

class_isMetaClass⚠

class_replaceMethod⚠

class_replaceProperty⚠

class_respondsToSelector⚠

class_setIvarLayout⚠

class_setVersion⚠

class_setWeakIvarLayout⚠

imp_getBlock⚠

imp_implementationWithBlock⚠

imp_removeBlock⚠

ivar_getName⚠

ivar_getOffset⚠

ivar_getTypeEncoding⚠

method_copyArgumentType⚠

method_copyReturnType⚠

method_exchangeImplementations⚠

method_getArgumentType⚠

method_getDescription⚠

method_getImplementation⚠

method_getName⚠

method_getNumberOfArguments⚠

method_getReturnType⚠

method_getTypeEncoding⚠

method_invoke⚠

method_invoke_stret⚠

method_setImplementation⚠

objc_addExceptionHandler⚠

objc_alloc⚠

objc_allocWithZone⚠

objc_allocateClassPair⚠

objc_allocateProtocol⚠

objc_autorelease⚠

objc_autoreleasePoolPop⚠

objc_autoreleasePoolPush⚠

objc_autoreleaseReturnValue⚠

objc_begin_catch⚠

objc_constructInstance⚠Deprecated

objc_copyClassList⚠

objc_copyClassNamesForImage⚠

objc_copyImageNames⚠

objc_copyProtocolList⚠

objc_copyWeak⚠

objc_destroyWeak⚠

objc_destructInstance⚠Deprecated

objc_disposeClassPair⚠

objc_duplicateClass⚠

objc_end_catch⚠

objc_enumerationMutation⚠

objc_exception_rethrow⚠

objc_exception_throw⚠

See objc-exception.h.

objc_exception_try_enter⚠

objc_exception_try_exit⚠

objc_getAssociatedObject⚠

objc_getClass⚠

objc_getClassList⚠

objc_getFutureClass⚠Deprecated

objc_getMetaClass⚠

objc_getProtocol⚠

objc_getRequiredClass⚠

objc_initWeak⚠

objc_loadWeak⚠

objc_loadWeakRetained⚠

objc_lookUpClass⚠

objc_moveWeak⚠

objc_msgSend⚠

objc_msgSendSuper⚠

objc_msgSendSuper_stret⚠

objc_msgSend_fp2ret⚠

objc_msgSend_fpret⚠

objc_msgSend_stret⚠

objc_msg_lookup⚠

objc_msg_lookup_stret⚠

objc_msg_lookup_super⚠

objc_msg_lookup_super_stret⚠

objc_registerClassPair⚠

objc_registerProtocol⚠

objc_release⚠

objc_removeAssociatedObjects⚠

objc_removeExceptionHandler⚠

objc_retain⚠

objc_retainAutorelease⚠

objc_retainAutoreleaseReturnValue⚠

objc_retainAutoreleasedReturnValue⚠

objc_retainBlock⚠

objc_setAssociatedObject⚠

objc_setEnumerationMutationHandler⚠

objc_setExceptionMatcher⚠

objc_setExceptionPreprocessor⚠

objc_setForwardHandler⚠

objc_setUncaughtExceptionHandler⚠

objc_storeStrong⚠

objc_storeWeak⚠

objc_sync_enter⚠

objc_sync_exit⚠

objc_terminate⚠

object_copy⚠Deprecated

object_dispose⚠Deprecated

object_getClass⚠

object_getClassName⚠

object_getIndexedIvars⚠

object_getInstanceVariable⚠Deprecated

object_getIvar⚠

object_isClass⚠

object_setClass⚠

object_setInstanceVariable⚠Deprecated

object_setIvar⚠

property_copyAttributeList⚠

property_copyAttributeValue⚠

property_getAttributes⚠

property_getName⚠

protocol_addMethodDescription⚠

protocol_addProperty⚠

protocol_addProtocol⚠

protocol_conformsToProtocol⚠

protocol_copyMethodDescriptionList⚠

protocol_copyPropertyList⚠

protocol_copyProtocolList⚠

protocol_getMethodDescription⚠

protocol_getName⚠

protocol_getProperty⚠

protocol_isEqual⚠

sel_getName⚠

sel_getUid⚠

sel_isEqual⚠

sel_isMapped⚠

sel_registerName⚠

Type Aliases

BOOL

The Objective-C BOOL type.

IMP

A nullable pointer to the start of a method implementation.

NSInteger

A signed integer value type.

NSUInteger

Describes an unsigned integer.

SEL

An immutable pointer to a selector.

id

A mutable pointer to an object / instance.

objc_AssociationPolicy

Policies related to associative references.

objc_exception_handler

Remember that this is non-null!

objc_exception_matcher

Remember that this is non-null!

objc_exception_preprocessor

Remember that this is non-null!

objc_uncaught_exception_handler

Remember that this is non-null!